Fix Issues Remove label
